requirements.txt

entsoe-py>=0.6.13
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
openpyxl>=3.1
brightway2>=2.4
python-dotenv>=1.0
requests>=2.31
